Reverted to non-rest ish way
