+++ src/main/java/com/iota/iri/service/restserver/ApiProcessor.java
@@ -1,0 +1,10 @@
+package com.iota.iri.service.restserver;
+
+import java.net.InetAddress;
+
+import com.iota.iri.service.dto.AbstractResponse;
+
+public interface ApiProcessor {
+    
+    AbstractResponse processFunction(String request, InetAddress inetAddress);
+}
