+++ src/main/java/com/iota/iri/service/restserver/RestConnector.java
@@ -1,12 +1,8 @@
 package com.iota.iri.service.restserver;
 
-import java.util.function.Function;
-
-import com.iota.iri.service.dto.AbstractResponse;
-
 public interface RestConnector {
 
-    void init(Function<String, AbstractResponse> processFunction);
+    void init(ApiProcessor processFunction);
 
     void start();
