+++ src/main/java/com/iota/iri/service/restserver/Root.java
@@ -3,11 +3,14 @@ package com.iota.iri.service.restserver;
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.ws.rs.Path;
 import javax.ws.rs.core.Application;
 
 public class Root extends Application {
     
+    public Root(ApiProcessor processFunction) {
+        
+    }
+
     @Override
     public Set<Class<?>> getClasses(){
        HashSet<Class<?>> classes = new HashSet<Class<?>>();
